+++ Tests/Editor/DucktionTestConfig.cs
@@ -1,0 +1,12 @@
+﻿namespace TheRealIronDuck.Ducktion.Editor.Tests.Editor
+{
+    public struct DucktionTestConfig
+    {
+        public readonly bool CreateContainer;
+
+        public DucktionTestConfig(bool createContainer)
+        {
+            CreateContainer = createContainer;
+        }
+    }
+}