+++ Tests/Editor/Stubs/RecursiveWrapperService.cs
@@ -1,4 +1,4 @@
-﻿namespace TheRealIronDuck.Ducktion.Tests.Stubs
+﻿namespace TheRealIronDuck.Ducktion.Editor.Tests.Editor.Stubs
 {
     /// <summary>
     /// Just a simple service which holds a reference to a recursive service.