+++ Tests/Editor/Stubs/SecondSimpleService.cs
@@ -1,4 +1,4 @@
-﻿namespace TheRealIronDuck.Ducktion.Tests.Stubs
+﻿namespace TheRealIronDuck.Ducktion.Editor.Tests.Editor.Stubs
 {
     /// <summary>
     /// This is just a simple service which also extends the ISimpleInterface