+++ Tests/Editor/Stubs/SimpleInterface.cs
@@ -1,0 +1,7 @@
+﻿namespace TheRealIronDuck.Ducktion.Editor.Tests.Editor.Stubs
+{
+    public interface ISimpleInterface
+    {
+        
+    }
+}