+++ Tests/Runtime/Stubs/SimpleInterface.cs
@@ -1,7 +1,0 @@
-﻿namespace TheRealIronDuck.Ducktion.Tests.Stubs
-{
-    public interface ISimpleInterface
-    {
-        
-    }
-}